Start avoiding floats in game logic.
The idea is to make the core deterministic, so that replays can be made
by simply recording RNG state at the start and player actions throughout
the game.

This will also enable Cool Things like verified leaderboards without
requiring people to play on a remote server.

NB I don't know if decimal will be up to the task, or whether it'll have
to be replaced by custom fixed point type, but at least it will mark
data that is supposed to be used in the game logic.

+++ LabyrinthCS/Data/MonsterData.cs
@@ -17,12 +17,12 @@ namespace Labyrinth.Data
         public Name Name { get; private set; }
         public string Description { get; private set; }
 
-        public float SpeedFactor { get; private set; }
+        public decimal SpeedFactor { get; private set; }
 
         public MonsterData(Id<Monster> id)
             : base(id)
         {
-            SpeedFactor = 1.0f;
+            SpeedFactor = 1.0m;
         }
 
         static MonsterData()
@@ -46,7 +46,7 @@ namespace Labyrinth.Data
                     // TODO flying
                     Name = new Name("bat"),
                     Description = "An unusually aggressive kind of bat.",
-                    SpeedFactor = 2.0f
+                    SpeedFactor = 2.0m
                 }
             };
         }
