+++ LabyrinthCS/Entities/Time/MoveCost.cs
@@ -4,7 +4,7 @@ namespace Labyrinth.Entities.Time
 {
     public static class MoveCost
     {
-        public static int ApplyFactor(float factor, int @base = Const.SpeedBase)
+        public static int ApplyFactor(decimal factor, int @base = Const.SpeedBase)
         {
             return MathExt.RoundInt(factor * @base);
         }
